Improve notification's messages.
