+++ client/src/main/kotlin/io/spine/examples/pingh/client/TimestampExts.kt
@@ -27,10 +27,59 @@
 package io.spine.examples.pingh.client
 
 import com.google.protobuf.Timestamp
-import com.google.protobuf.Duration
 import com.google.protobuf.util.Timestamps
+import java.time.Duration
+import java.time.LocalDateTime
+import java.time.ZoneOffset
+import java.time.format.DateTimeFormatter
+
+/**
+ * A time format consisting of the day of the month and the name of the month.
+ */
+private val dateFormat = DateTimeFormatter.ofPattern("dd MMM")
+
+/**
+ * Number of minutes in one hour.
+ */
+private const val minutesPerHour = 60L
+
+/**
+ * Number of hours in one day.
+ */
+private const val hoursPerDay = 24L
 
 /**
  * Adds a duration to this timestamp.
  */
-internal fun Timestamp.add(duration: Duration): Timestamp = Timestamps.add(this, duration)
+internal fun Timestamp.add(duration: com.google.protobuf.Duration): Timestamp =
+    Timestamps.add(this, duration)
+
+/**
+ * Converts a `Timestamp` to a string based on the time difference between
+ * the given past timestamp and the current moment.
+ *
+ * - If the difference is less than a minute, returns the string `"just now"`;
+ * - If the difference is between one minute and one hour, returns the difference in minutes;
+ * - If the difference is between one hour and one day, returns the difference in hours;
+ * - If the difference is exactly one day, returns the string `"yesterday"`;
+ * - In all other cases, returns the day and month.
+ *
+ * @throws IllegalArgumentException if the `Timestamp` is not from the past.
+ */
+public fun Timestamp.howMuchTimeHasPassed(): String {
+    val thisDatetime = LocalDateTime.ofEpochSecond(this.seconds, this.nanos, ZoneOffset.UTC)
+    val difference = Duration.between(
+        thisDatetime,
+        LocalDateTime.now(ZoneOffset.UTC)
+    )
+    require(difference > Duration.ZERO) { "" }
+    return when {
+        difference.toMinutes() < 1L -> "just now"
+        difference.toMinutes() == 1L -> "a minute ago"
+        difference.toMinutes() < minutesPerHour -> "${difference.toMinutes()} minutes ago"
+        difference.toHours() == 1L -> "an hour ago"
+        difference.toHours() < hoursPerDay -> "${difference.toHours()} hours ago"
+        difference.toDays() == 1L -> "yesterday"
+        else -> dateFormat.format(thisDatetime)
+    }
+}
